+++ tests/Feature/Form/FormField/FormFieldErrors.php
@@ -2,6 +2,7 @@
 
 namespace Portavice\Bladestrap\Tests\Feature\Form\FormField;
 
+use Portavice\Bladestrap\Support\Options;
 use Portavice\Bladestrap\Tests\Feature\ComponentTestCase;
 
 class FormFieldErrors extends ComponentTestCase
@@ -35,13 +36,7 @@ class FormFieldErrors extends ComponentTestCase
     public static function allTypes(): array
     {
         return [
-            ...array_map(
-                static fn ($value) => [
-                    $value[0],
-                    $value[1] . "\n" . 'ERROR_PLACEHOLDER',
-                ],
-                self::types()
-            ),
+            // type checkbox with single option, given as array or via Options helper
             [
                 'type="checkbox" :options="[1 => 1]"',
                 '<div class="form-check">
@@ -50,26 +45,52 @@ class FormFieldErrors extends ComponentTestCase
                     ERROR_PLACEHOLDER
                 </div>',
             ],
+            [
+                'type="checkbox" :options="\Portavice\Bladestrap\Support\Options::one(1)"',
+                '<div class="form-check">
+                    <input id="test-1" name="test" type="checkbox" value="1" class="form-check-input is-invalid"/>
+                    <label class="form-check-label" for="test-1">1</label>
+                    ERROR_PLACEHOLDER
+                </div>',
+            ],
+
+            // type checkbox with multiple options, given as array or via Options helper
+            self::makeDataForTypeAndOptions('checkbox', '[1 => 1, 2 => 2]'),
+            self::makeDataForTypeAndOptions('checkbox', '\Portavice\Bladestrap\Support\Options::fromArray([1 => 1, 2 => 2])'),
+
+            // type radio with multiple options, given as array or via Options helper
+            self::makeDataForTypeAndOptions('radio', '[1 => 1, 2 => 2]'),
+            self::makeDataForTypeAndOptions('radio', '\Portavice\Bladestrap\Support\Options::fromArray([1 => 1, 2 => 2])'),
+
+            // type range, select, text, textarea
             ...array_map(
-                static fn ($type) => [
-                    'type="' . $type . '" :options="[1 => 1, 2 => 2]"',
-                    '<div class="form-check">
-                        <input id="test-1" name="test" type="' . $type . '" value="1" class="form-check-input is-invalid"/>
-                        <label class="form-check-label" for="test-1">1</label>
-                    </div>
-                    <div class="form-check">
-                        <input id="test-2" name="test" type="' . $type . '" value="2" class="form-check-input is-invalid"/>
-                        <label class="form-check-label" for="test-2">2</label>
-                        ERROR_PLACEHOLDER
-                    </div>',
+                static fn ($value) => [
+                    $value[0],
+                    $value[1] . "\n" . 'ERROR_PLACEHOLDER',
                 ],
-                ['checkbox', 'radio']
+                self::typesSupportingInputGroups()
             ),
         ];
     }
 
+    private static function makeDataForTypeAndOptions(string $type, string $options): array
+    {
+        return [
+            'type="' . $type . '" :options="' . $options . '"',
+            '<div class="form-check">
+                <input id="test-1" name="test" type="' . $type . '" value="1" class="form-check-input is-invalid"/>
+                <label class="form-check-label" for="test-1">1</label>
+            </div>
+            <div class="form-check">
+                <input id="test-2" name="test" type="' . $type . '" value="2" class="form-check-input is-invalid"/>
+                <label class="form-check-label" for="test-2">2</label>
+                ERROR_PLACEHOLDER
+            </div>',
+        ];
+    }
+
     /**
-     * @dataProvider types
+     * @dataProvider typesSupportingInputGroups
      */
     public function testFormFieldWithInputGroupShowsErrors(string $type, string $output): void
     {
@@ -96,7 +117,7 @@ class FormFieldErrors extends ComponentTestCase
         );
     }
 
-    public static function types(): array
+    public static function typesSupportingInputGroups(): array
     {
         return [
             [