+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Display of error messages for radios and checkboxes whose name is not an array
+
 
 ## Version 1.1.0 (2023-12-30)