+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed
-- Display of error messages for radios and checkboxes whose name is not an array
+- Display of error messages for checkboxes and radios whose name is not an array
+- Make `Portavice\Bladestrap\Support\Options` countable to correct `$loop->last` in Blade
+  and display of error messages for checkboxes and radios
 
 
 ## Version 1.1.0 (2023-12-30)
@@ -33,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Components for [dropdown](https://getbootstrap.com/docs/5.3/components/dropdowns/#single-button) buttons (with different directions), headers, and items
 - Components for [forms](https://getbootstrap.com/docs/5.3/forms/overview/) and their fields
   - Support all types of `<input>`s, `<textarea>`, and `<select>`
-  - Set HTML attributes (e.g. classes or data attributes) for options of `<input type="checkbox">`, `<input type="radio">`, and `<select>`:
+  - Set HTML attributes (e.g. classes or data attributes) for options of `<input type="checkbox">`, `<input type="radio">`, and `<select>`
   - Allow HTML for labels of `<input type="checkbox">` and `<input type="radio">`
   - Prefill values with old values or from query parameters
   - Show feedback in case of errors